test: restructure barbacane-test into per-feature integration test files
The monolithic 4200-line `mod tests` block in `src/gateway.rs` is split
into 15 focused files under `crates/barbacane-test/tests/`:

  routing, validation, proxy, auth, rate_limit, cache, telemetry,
  lifecycle, events, specs, plugins, compilation, access_control,
  admin, streaming

Each file is a standard Cargo integration test (top-level #[tokio::test]
functions, no mod wrapper) and uses a local `fixture()` helper to resolve
paths relative to the workspace root.

Also adds:
- `tests/fixture-plugins/streaming-echo`: a minimal WASM dispatcher that
  calls `host_http_stream` and returns `streamed_response()`, used by
  the streaming integration tests.
- `crates/barbacane-test/build.rs`: auto-compiles fixture plugins before
  the test binary links; emits `cargo:warning` on failure so the build
  does not hard-fail when `wasm32-unknown-unknown` is unavailable.
- `scripts/count-tests.sh`: updated grep pattern — integration tests are
  now top-level (no `gateway::` prefix) in separate test binaries.
