+++ crates/barbacane-test/src/lib.rs
@@ -1,10 +1,11 @@
 //! Test harnesses for Barbacane gateway and plugins.
 //!
-//! Provides `TestGateway` for full-stack integration tests
-//! and `PluginHarness` for isolated WASM plugin testing.
+//! Provides `TestGateway` for full-stack integration tests.
 
 #[cfg(test)]
 pub mod cli;
 pub mod gateway;
 
-pub use gateway::{TestError, TestGateway};
+pub use gateway::{
+    assert_status, generate_test_certificates, TestCertificates, TestError, TestGateway,
+};
